+++ lib/Traits/TraitsClasses/AssImpCatalogItemTraits.ts
@@ -3,7 +3,7 @@ import primitiveTrait from "../Decorators/primitiveTrait";
 import mixTraits from "../mixTraits";
 import GltfCatalogItemTraits from "./GltfCatalogItemTraits";
 
-export default class ColladaCatalogItemTraits extends mixTraits(
+export default class AssImpCatalogItemTraits extends mixTraits(
   GltfCatalogItemTraits
 ) {
   @primitiveArrayTrait({
@@ -19,4 +19,11 @@ export default class ColladaCatalogItemTraits extends mixTraits(
     description: `The base URL that paths in the 3D model (eg textures) are relative to`
   })
   baseUrl?: string;
+
+  @primitiveTrait({
+    type: "boolean",
+    name: "Debug download GLTF",
+    description: `If true, then GLTF output files will be automatically downloaded`
+  })
+  debugDownloadGltf?: boolean;
 }