+++ CHANGES.md
@@ -3,6 +3,10 @@ Change Log
 
 #### next release (8.2.7)
 
+* Add `GlTf` type definition (v2)
+* Add experimental client-side 3D file conversion using [`assimpjs`](https://github.com/kovacsv/assimpjs) ([emscripten](https://emscripten.org) interface for the [assimp](https://github.com/assimp/assimp) library)
+  * This supports `zip` files through add local/remote data
+  * Supports over 40 formats - including Collada, obj, Blender, DXF - [full list](https://github.com/assimp/assimp/blob/master/doc/Fileformats.md)
 * [The next improvement]
 
 #### 8.2.6 - 2022-06-17